code/requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-huggingface>=0.0.3
sentence-transformers>=3.0.0
langchain-community>=0.3.0
faiss-cpu>=1.9.0
rich>=13.0.0
python-dotenv>=1.0.0
tiktoken>=0.8.0
